Merging from Upstream

+++ src/StaticCaching/Cachers/Writer.php
@@ -60,6 +60,10 @@ class Writer
 
     public function flush($path)
     {
+        if (! File::exists($path)) {
+            return;
+        }
+
         foreach (Folder::getFilesRecursively($path) as $file) {
             File::delete($file);
         }
